@@ -8,27 +8,24 @@
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "api": "ng-openapi-gen",
-    "dev:ssr": "ng run client:serve-ssr",
-    "serve:ssr": "node dist/client/server/main.js",
-    "build:ssr": "ng build && ng run client:server",
-    "prerender": "ng run client:prerender",
-    "serve:ssr:@plhx/xx": "node dist/plhx/xx/server/server.mjs"
+    "serve:ssr:@plhx/xx": "node dist/plhx/xx/server/server.mjs",
+    "serve:ssr:client": "node dist/client/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.8",
-    "@angular/cdk": "17.0.4",
-    "@angular/common": "^17.0.8",
-    "@angular/compiler": "^17.0.8",
-    "@angular/core": "^17.0.8",
-    "@angular/forms": "^17.0.8",
-    "@angular/material": "17.0.4",
-    "@angular/platform-browser": "^17.0.8",
-    "@angular/platform-browser-dynamic": "^17.0.8",
-    "@angular/platform-server": "^17.0.8",
-    "@angular/router": "^17.0.8",
-    "@angular/service-worker": "^17.0.8",
-    "@angular/ssr": "^17.0.8",
+    "@angular/animations": "^18.0.1",
+    "@angular/cdk": "18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "^18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "^18.0.1",
+    "@angular/material": "18.0.1",
+    "@angular/platform-browser": "^18.0.1",
+    "@angular/platform-browser-dynamic": "^18.0.1",
+    "@angular/platform-server": "^18.0.1",
+    "@angular/router": "^18.0.1",
+    "@angular/service-worker": "^18.0.1",
+    "@angular/ssr": "^18.0.2",
     "@fortawesome/angular-fontawesome": " 0.13.0",
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
     "@fortawesome/free-brands-svg-icons": "^6.4.0",
@@ -37,6 +34,9 @@
     "@jonyadamit/ngx-qrcode-ivy": "^16.0.0",
     "@swimlane/ngx-charts": "^20.4.1",
     "crypto-js": "^4.1.1",
+    "d3-scale": "^4.0.2",
+    "d3-selection": "^3.0.0",
+    "d3-shape": "^3.2.0",
     "express": "^4.18.2",
     "firebase": "^10.8.0",
     "rxjs": "~7.8.0",
@@ -44,10 +44,13 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.8",
-    "@angular/cli": "~17.0.8",
-    "@angular/compiler-cli": "^17.0.8",
+    "@angular-devkit/build-angular": "^18.0.2",
+    "@angular/cli": "~18.0.2",
+    "@angular/compiler-cli": "^18.0.1",
     "@types/crypto-js": "^4.1.1",
+    "@types/d3-scale": "^4.0.8",
+    "@types/d3-selection": "^3.0.10",
+    "@types/d3-shape": "^3.1.6",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~4.3.0",
     "@types/node": "^18.18.0",
@@ -58,8 +61,8 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
     "monaco-editor": "^0.44.0",
-    "ng-packagr": "^17.0.0",
+    "ng-packagr": "^18.0.0",
     "ngx-monaco-editor": "^12.0.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.4.5"
   }
 }
