@@ -1,6 +1,6 @@
 import { ModuleWithProviders, NgModule, isDevMode } from "@angular/core";
 import { CommonModule } from "@angular/common";
-import { HTTP_INTERCEPTORS, HttpClientModule } from "@angular/common/http";
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi, HttpClientModule } from "@angular/common/http";
 import { ApiModule } from "@portal/api";
 import { AutoFocusDirective } from "./auto-focus.directive";
 import { StorageService } from "./storage.service";
@@ -11,25 +11,23 @@ import { DaoService } from "./dao/dao.service";
 import { ToolbarService } from "./services/toolbar.service";
 import { NotificationService } from "./services/notification.service";
 import { ServicesService } from "./services/services.service";
+import { LocalizationService } from "./services/localization.service";
+import { TranslateDirective } from "./i18n/translate.directive";
 
-@NgModule({
-    imports: [
-        CommonModule,
-        HttpClientModule,
-    ],
-    declarations: [
+@NgModule({ declarations: [
         PurePipe,
         AutoFocusDirective,
-    ],
-    providers: [
     ],
     exports: [
         PurePipe,
         CommonModule,
         HttpClientModule,
         AutoFocusDirective,
-    ]
-})
+        TranslateDirective,
+    ], imports: [CommonModule,
+        TranslateDirective], providers: [
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class CoreModule {
     public static forRoot(): ModuleWithProviders<CoreModule> {
         return {
@@ -45,6 +43,7 @@ export class CoreModule {
                 ToolbarService,
                 NotificationService,
                 ServicesService,
+                LocalizationService,
                 {
                     provide: HTTP_INTERCEPTORS, multi: true, useClass: AuthorizationHttpInterceptor
                 },
