@@ -1,9 +1,10 @@
-import { Component, Inject, isDevMode } from '@angular/core';
-import { ChildActivationStart, Router } from '@angular/router';
+import { Component, Inject, Optional, isDevMode } from '@angular/core';
+import { ChildActivationStart, NavigationError, Router } from '@angular/router';
+import { DOCUMENT } from '@angular/common';
 import { map } from 'rxjs';
 import { UserService } from './services/user.service';
 import { PaginaErroComponent } from './views/pagina-erro/pagina-erro.component';
-import { DOCUMENT } from '@angular/common';
+import { LocalizationService } from './core/services/localization.service';
 @Component({
   selector: 'ci-root',
   templateUrl: './app.component.html',
@@ -16,9 +17,16 @@ export class AppComponent {
     private readonly userService: UserService,
     @Inject(DOCUMENT)
     private readonly document: Document,
+    @Optional()
+    localizacao: LocalizationService,
   ) {
+    setTimeout(() => localizacao.init(), 0);
     router.events.subscribe(v => {
       try {
+        if (v instanceof NavigationError) {
+          router.navigate(['/']);
+        }
+        console.log(v.constructor.name);
         if (!!(v as any)?.snapshot?.data?.name) {
           document.title =
             `${isDevMode() ? '[dev] ' : ''}` +
