@@ -1,16 +1,16 @@
 import { NgModule, isDevMode } from '@angular/core';
 import { BrowserModule, provideClientHydration } from '@angular/platform-browser';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ServiceWorkerModule } from '@angular/service-worker';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { MonacoEditorModule } from 'ngx-monaco-editor';
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
 import { ServicesModule } from './services/services.module';
 import { CoreModule } from './core/core.module';
 import { ContextmenuModule } from './components/contextmenu/contextmenu.module';
-import { MonacoEditorModule } from 'ngx-monaco-editor';
-import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { PaginaErroComponent } from './views/pagina-erro/pagina-erro.component';
+import { OrganizacaoService } from './services/organizacao.service';
 
 @NgModule({
   declarations: [
@@ -34,7 +34,8 @@ import { PaginaErroComponent } from './views/pagina-erro/pagina-erro.component';
     FontAwesomeModule, // use forRoot() in main app module only.
   ],
   providers: [
-    provideClientHydration()
+    provideClientHydration(),
+    OrganizacaoService,
   ],
   bootstrap: [
     AppComponent,
