@@ -1,4 +1,6 @@
-import { Component } from '@angular/core';
+import { Component, Inject, Input, Optional } from '@angular/core';
+import { FORM_OPTIONS, IFormOptions } from './i-form-options';
+import { FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'ci-dyn-form',
@@ -6,5 +8,14 @@ import { Component } from '@angular/core';
   styleUrls: ['./dyn-form.component.scss']
 })
 export class DynFormComponent {
-
+  @Input()
+  formGroup?: FormGroup;
+  @Input()
+  source?: IFormOptions;
+  constructor(
+    @Optional() @Inject(FORM_OPTIONS)
+    formOptions?: IFormOptions,
+  ) {
+    if (formOptions) this.source = formOptions;
+  }
 }
