@@ -1,12 +1,16 @@
-import { Component, ElementRef, HostListener } from '@angular/core';
-import { ActivatedRoute, ActivatedRouteSnapshot, ActivationEnd, ActivationStart, ChildActivationEnd, Router } from '@angular/router';
+import { Component, HostListener, Optional } from '@angular/core';
+import { ActivatedRouteSnapshot, ActivationEnd, ChildActivationEnd, Router } from '@angular/router';
 import { UserService } from 'src/app/services/user.service';
 import { lastValueFrom } from 'rxjs';
-import { Application } from '@portal/api';
+import { Application, CpuInfo, SystemService, Tenant } from '@portal/api';
 import { ApplicationService } from '@portal/api';
-import { HttpClient } from '@angular/common/http';
 import { ServicesService } from 'src/app/core/services/services.service';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
+import { OrganizacaoService } from 'src/app/services/organizacao.service';
+import { WindowService } from '../window/window.service';
+import { SettingsComponent } from 'src/app/views/settings/settings.component';
+// import { WindowService } from '../window/window.service';
+// import { SettingsComponent } from 'src/app/views/settings/settings.component';
 
 interface IBreadcrumb {
   name?: string;
@@ -23,6 +27,8 @@ export class LNavComponent {
   title?: string;
   breadcrumb?: IBreadcrumb[];
   $user = this.userService.user;
+  tenants?: Tenant[];
+  cpuStatus?: any[];
   get appsGlobal() { return this.apps?.filter(i => i.menuGroupName === 'global') }
   get appsUser() { return this.apps?.filter(i => i.menuGroupName === 'user') }
   userPhoto?: SafeResourceUrl;
@@ -30,9 +36,15 @@ export class LNavComponent {
     private readonly router: Router,
     private readonly userService: UserService,
     private readonly applicationService: ApplicationService,
-    private readonly http: HttpClient,
+    /// private readonly http: HttpClient,
     public readonly services: ServicesService,
     private readonly sanitizer: DomSanitizer,
+    private readonly system: SystemService,
+    @Optional()
+    private readonly organizacaoService?: OrganizacaoService,
+
+
+    private readonly window?: WindowService,
   ) {
     this.load();
     router.events.subscribe((event: any) => {
@@ -41,7 +53,6 @@ export class LNavComponent {
           event instanceof ActivationEnd ||
           event instanceof ChildActivationEnd
         )
-
         && event && event.snapshot instanceof ActivatedRouteSnapshot &&
         event?.snapshot?.data) {
         const data: { name?: string, icon?: string } = event.snapshot.data;
@@ -59,10 +70,65 @@ export class LNavComponent {
       }
     });
     userService.user.subscribe(user => {
-      if (user?.photo?.originalFile)
-        this.userPhoto = this.sanitizer.bypassSecurityTrustResourceUrl('data:image;base64,' + btoa(String.fromCharCode(...new Uint8Array((user?.photo?.originalFile as any).data))));
+      if (user?.photo?.originalFile) {
+        const u8 = new Uint8Array((user?.photo?.originalFile as any).data);
+        this.userPhoto = this.sanitizer.bypassSecurityTrustResourceUrl('data:image/jpeg;base64,' +
+          btoa(this.Uint8ToString(u8))
+        );
+      }
     })
   }
+  private async updateCpuInfo() {
+    try {
+      const cpu_infos = (await lastValueFrom(this.system.systemLeitura()));
+      const x: any = {};
+      x['user'] = { name: 'User', series: [] };
+      x['system'] = { name: 'System', series: [] };
+      x['heapTotal'] = { name: 'heapTotal', series: [] };
+      x['heapUsed'] = { name: 'heapUsed', series: [] };
+      x['rss'] = { name: 'rss', series: [] };
+      x['external'] = { name: 'external', series: [] };
+      cpu_infos.forEach(i => {
+        x['user'].series.push({
+          "value": i.user,
+          "name": i.moment,
+        },);
+        x['system'].series.push({
+          "value": i.system,
+          "name": i.moment,
+        },);
+        x['heapTotal'].series.push({
+          "value": i.heapTotal,
+          "name": i.moment,
+        },);
+        x['heapUsed'].series.push({
+          "value": i.heapUsed,
+          "name": i.moment,
+        },);
+        x['rss'].series.push({
+          "value": i.rss,
+          "name": i.moment,
+        },);
+        x['external'].series.push({
+          "value": i.external,
+          "name": i.moment,
+        },);
+      });
+      this.cpuStatus = [...Object.values(x)];
+    } catch (error) {
+      console.error(error);
+    }
+    setTimeout(() => this.updateCpuInfo(), 1000);
+  }
+  private Uint8ToString(u8a: any) {
+    var CHUNK_SZ = 0x8000;
+    var c = [];
+    for (var i = 0; i < u8a.length; i += CHUNK_SZ) {
+      c.push(String.fromCharCode.apply(null, u8a.subarray(i, i + CHUNK_SZ)));
+    }
+    return c.join("");
+  }
+
   load() {
     (async () => {
       //try {
@@ -73,6 +139,8 @@ export class LNavComponent {
       //}
     })();
     this.loadStatus();
+    if (this.$user.value?.tenants) this.tenants = this.$user.value?.tenants;
+    // this.updateCpuInfo();
   }
   sair() {
     this.userService.sair();
@@ -112,5 +180,8 @@ export class LNavComponent {
   }
   action(crumb: any) {
     console.log(crumb);
+  }
+  async settings() {
+    this.window?.open(SettingsComponent, null);
   }
 }
