@@ -1,4 +1,9 @@
 import { Component } from '@angular/core';
+import { InfraService } from '../services/infra.service';
+import { Domain, DomainService } from '@portal/api';
+import { WindowService } from 'src/app/components/window/window.service';
+import { DomainComponent } from './casdastros/domain/domain.component';
+import { DaoService } from 'src/app/core/dao/dao.service';
 
 @Component({
   selector: 'ci-inicio',
@@ -6,5 +11,9 @@ import { Component } from '@angular/core';
   styleUrls: ['./inicio.component.scss']
 })
 export class InicioComponent {
-
+  constructor(
+    public readonly service: InfraService,
+  ) { }
+  addDomain() {
+  }
 }
