@@ -1,4 +1,7 @@
 import { NgModule } from "@angular/core";
+import { RouterModule } from "@angular/router";
+import { MasterDetailComponent } from "src/app/components/master-detail/master-detail.component";
+import { MasterDetailModule } from "src/app/components/master-detail/master-detail.module";
 
 
 /**
@@ -6,6 +9,14 @@ import { NgModule } from "@angular/core";
  * 
  */
 @NgModule({
-
+    imports: [
+        MasterDetailModule,
+        RouterModule.forChild([
+            {
+                path: '',
+                component: MasterDetailComponent
+            }
+        ])
+    ]
 })
 export class LowCodeModule { }
