@@ -9,14 +9,18 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatMenuModule } from '@angular/material/menu';
-import { AdicionarWidgetComponent } from './inicio/adicionar-widget/adicionar-widget.component';
-import { WindowModule } from 'src/app/components/window/window.module';
-import { LayoutComponent } from './inicio/layout/layout.component';
-import { LayoutThumbComponent } from './inicio/layout/layout-thumb/layout-thumb.component';
+import { MatTabsModule } from '@angular/material/tabs';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
 import { WidgetService } from './inicio/widget.service';
+import { AdicionarWidgetComponent } from './inicio/adicionar-widget/adicionar-widget.component';
+import { WindowModule } from 'src/app/components/window/window.module';
+import { LayoutThumbComponent } from './inicio/layout/layout-thumb/layout-thumb.component';
+import { LayoutComponent } from './inicio/layout/layout.component';
+import { PranchetaService } from './config.service';
+import { PranchetaComponent } from './inicio/prancheta/prancheta.component';
 
 @NgModule({
   declarations: [
@@ -39,9 +43,15 @@ import { WidgetService } from './inicio/widget.service';
     MatInputModule,
     MatMenuModule,
     ReactiveFormsModule,
+    MatTabsModule,
+    FormsModule,
+    ReactiveFormsModule,
+    PranchetaComponent,
+    
   ],
   providers: [
     WidgetService,
+    PranchetaService,
   ]
 })
 export class PainelModule { }
