@@ -35,7 +35,9 @@ export class AuthGuard implements CanActivate {
       });
       // 💡 We're assigning the payload to the request object here
       // so that we can access it in our route handlers
-      request['user'] = payload;
+      console.log('PAYLOAD', payload)
+      request['user'] = { id: payload.id };
+      request['chaveAcesso'] = payload.chaveAcesso;
     } catch {
       throw new UnauthorizedException();
     }
