@@ -8,18 +8,22 @@ import { User } from "./models/user.entity";
 import { jwtConstants } from "./constants";
 import { AuthGuard } from "./auth.guard";
 import { CredencialService } from "./service/credencial.service";
-import { ChaveAcesso } from "./models/chave-acesso.entity";
+import { ChaveAcesso } from "../core/audt/chave-acesso.entity";
 import { Policy } from "./models/policy.entity";
 import { RefreshTokenStrategy } from "./service/refresh-token-strategy";
 import { AuthService } from "./service/auth.service";
 import { DeviceAuthenticated } from "./models/device-autenticated.entity";
 import { TenantModule } from "src/tenant/tenant.module";
+import { AccessCredential } from "./models/user-credential.entity";
+import { UserCredentialService } from "./service/user-credential.service";
+import { CoreModule } from "src/core/core.module";
 
 export const AuthEntities = [
     Policy,
     User,
     ChaveAcesso,
     DeviceAuthenticated,
+    AccessCredential,
 ];
 
 @Module({
@@ -33,6 +37,7 @@ export const AuthEntities = [
             signOptions: { expiresIn: '60s' },
         }),
         TenantModule,
+        CoreModule,
     ],
     controllers: [
         AuthController,
@@ -41,6 +46,7 @@ export const AuthEntities = [
         UserService,
         AuthService,
         CredencialService,
+        UserCredentialService,
         RefreshTokenStrategy,
         {
             provide: APP_GUARD,
@@ -49,6 +55,7 @@ export const AuthEntities = [
     ],
     exports: [
         UserService,
+        AuthService,
     ]
 })
 export class AuthModule { }
