@@ -1,5 +1,5 @@
 import { FullAuditedEntity } from "src/core/dao";
-import { Column, Entity, JoinTable, ManyToMany } from "typeorm";
+import { Column, Entity, JoinTable, ManyToMany, ManyToOne, OneToMany } from "typeorm";
 import { Endereco } from "./endereco.entity";
 import { ApiProperty } from "@nestjs/swagger";
 import { InformacaoContato } from "./informacao-contato.entity";
@@ -16,25 +16,25 @@ export class Pessoa extends FullAuditedEntity {
      * Nome
      */
     @ApiProperty({ required: false, nullable: true, description: 'Nome' })
-    @Column({ nullable: true })
+    @Column({ nullable: true, length: 120 })
     nome?: string;
     /**
      * Sobrenome
      */
     @ApiProperty({ required: false, nullable: true, description: 'Sobrenome' })
-    @Column({ nullable: true })
+    @Column({ nullable: true, length: 120 })
     sobrenome?: string;
     /**
      * Nome de Batismo ou Preferido em Origem
      */
-    @ApiProperty()
-    @Column({})
+    @ApiProperty({ nullable: true, required: false })
+    @Column({ length: 120, nullable: true })
     razaoSocial?: string;
     /**
      * Nome de Apresentação em Documentos Vinculados
      */
-    @ApiProperty()
-    @Column({})
+    @ApiProperty({ nullable: true, required: false })
+    @Column({ length: 120, nullable: true })
     nomeFantasia?: string;
     /**
      * Registro Geral em Caso de Pessoa Física registrada em território Brasileiro de acordo com a Constituição Federal. Obrigatório em casos de recolhimentos automatizados de documentos juntos ao estado. Sendo opcional para casos de alimentação manual de base. Esse documento se torna obrigatório em caso de automações junto ao estado em nome do próprio requerente. Sendo obrigatório a autorização direta do uso de seus dados. Com cancelamento ativo por parte do sistema em contato direto com o solicitante. 
@@ -54,7 +54,10 @@ export class Pessoa extends FullAuditedEntity {
         nullable: true
     })
     registroGeralRepublicaBrasileiraOrgaoEmissorOrgaoEmissor?: string;
-    @ApiProperty()
+    @ApiProperty({
+        nullable: true,
+        required: false,
+    })
     @Column({ nullable: true, length: 512 })
     emailPessoal?: string;
     @ApiProperty({ required: false, nullable: true })
@@ -64,7 +67,10 @@ export class Pessoa extends FullAuditedEntity {
     @ManyToMany(() => Endereco)
     @JoinTable()
     endereco?: Endereco[];
-    @ApiProperty()
+    @ApiProperty({
+        nullable: true,
+        required: false,
+    })
     @ManyToMany(() => InformacaoContato)
     @JoinTable()
     informacoesContato?: InformacaoContato[];
@@ -74,8 +80,13 @@ export class Pessoa extends FullAuditedEntity {
     @ApiProperty({ nullable: true, required: false })
     @Column({ nullable: true, enum: ['F', 'J'], length: 1 })
     tipoJuridico?: string;
-
-    @ApiProperty()
-    @ManyToMany(() => DocumentoIdentificacao)
-    documentos: DocumentoIdentificacao;
+    @ApiProperty({
+        nullable: true,
+        required: false,
+        type: DocumentoIdentificacao,
+        isArray: true,
+    })
+    @OneToMany(() => DocumentoIdentificacao, documento => documento.pessoa)
+    @JoinTable()
+    documentos?: DocumentoIdentificacao[];
 }
