@@ -1,23 +1,26 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { Agent } from "./models/agent";
+import { CodexService } from "./service/codex.service";
+import { CodexController } from "./controller/codex.controller";
+import { CodeBlock } from "./models/code-block.entity";
+import { CoreModule } from "src/core/core.module";
 export const CodeXEntities = [
     Agent,
+    CodeBlock,
 ];
 @Module({
     imports: [
+        CoreModule,
         TypeOrmModule.forFeature([
             ...CodeXEntities,
-        ])
-    ],
-    exports: [
-
+        ]),
     ],
     controllers: [
-
+        CodexController,
     ],
     providers: [
-
+        CodexService,
     ],
 })
 export class CodexModule { }
