@@ -1,6 +1,8 @@
 import { Entity, PrimaryColumn } from "typeorm";
 
-@Entity()
+@Entity({
+    schema: 'sys'
+})
 export class System {
     @PrimaryColumn({ type: 'varchar', length: 12 })
     mac?: string;
