@@ -1,17 +1,17 @@
 import { Module } from "@nestjs/common";
+import { TypeOrmModule } from "@nestjs/typeorm";
 import { SystemService } from "./system.service";
 import { SystemController } from "./system.controller";
 import { System } from "./model/system.entity";
+import { Status } from "./model/status";
 
 export const SystemEntities = [
+    Status,
     System,
 ];
 @Module({
     imports: [
-
-    ],
-    exports: [
-        
+        TypeOrmModule.forFeature([...SystemEntities])
     ],
     controllers: [
         SystemController,
