@@ -1,4 +1,18 @@
+import { ApiProperty } from "@nestjs/swagger";
 import { FullAuditedEntity } from "src/core/dao";
-import { Entity } from "typeorm";
+import { Column, Entity } from "typeorm";
 @Entity()
-export class Marca extends FullAuditedEntity { }
+export class Marca extends FullAuditedEntity {
+    @ApiProperty({ nullable: true, required: false })
+    @Column({ nullable: true })
+    name?: string;
+    @ApiProperty({ nullable: true, required: false })
+    @Column({ nullable: true, unique: true })
+    code?: string;
+    @ApiProperty({ nullable: true, required: false })
+    @Column({ nullable: true })
+    description?: string;
+    @ApiProperty({ nullable: true, required: false })
+    @Column({ nullable: true })
+    siteUrl?: string;
+}
