@@ -1,4 +1,4 @@
 import { FullAuditedEntity } from "src/core/dao";
 import { Entity } from "typeorm";
 @Entity()
-export class Marca extends FullAuditedEntity { }
+export class Patente extends FullAuditedEntity { }
