@@ -1,21 +1,36 @@
-import { Controller, Post } from "@nestjs/common";
+import { Body, Controller, Post } from "@nestjs/common";
 import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 import { DevicePayload } from "./dto/device-payload";
 import { DeviceService } from "../services/device.service";
+import { Public } from "src/auth/decorators/public.decorator";
+import { PoolDto } from "./dto/pool.dto";
 
 @Controller('Device')
 @ApiTags('Device')
 export class DeviceController {
     constructor(
-        private readonly devceService: DeviceService,
+        private readonly deviceService: DeviceService,
     ) { }
 
-    @ApiOperation({ operationId: 'Connect' })
+    @Public()
+    @ApiOperation({ operationId: 'DeviceConnect' })
     @Post('Connect')
     @ApiResponse({ type: DevicePayload })
-    async connectDevice(input: DevicePayload) {
-        return this.devceService.connect({
-            mac: 'not undetached'
+    public async connectDevice(@Body() input: DevicePayload) {
+        console.log(input);
+        return await this.deviceService.connect({
+            mac: input.id,
+            type: input.name
         });
+    }
+
+    @Public()
+    @ApiOperation({ operationId: 'DevicePool' })
+    @Post('Pool')
+    @ApiResponse({ type: PoolDto })
+    public async Pool(@Body() input: DevicePayload) {
+        let pool = new PoolDto();
+        pool.messages = [];
+        return await pool;
     }
 }
