@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { Repository } from "typeorm";
+import { Equal, Repository } from "typeorm";
 import { Device } from "../models/device.entity";
 import { InjectRepository } from "@nestjs/typeorm";
 
@@ -10,14 +10,19 @@ export class DeviceService {
         private readonly repo: Repository<Device>,
     ) { }
     async connect(device?: Device) {
-        if (device?.id) {
-            await this.createAndSave(device)
-        }
-        return device;
+        return await this.createAndSave(device)
     }
     async createAndSave(device?: Device) {
         if (!device) return;
+        const device_found = await this.find(device);
+        if (device_found) return device_found;
         device = this.repo.create(device);
         return await this.repo.save(device);
+    }
+    async find(device?: Device) {
+        if (!device) return;
+        if (device.mac)
+            return await this.repo.findOne({ where: { mac: Equal(device.mac) } });
+        return null;
     }
 }
