@@ -1,11 +1,12 @@
 import { ApiProperty } from "@nestjs/swagger";
+import { FullAuditedEntity } from "src/core/dao";
 import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
 
 @Entity()
-export class Photo {
-    @ApiProperty({ nullable: true, required: false })
-    @PrimaryGeneratedColumn('uuid')
-    id?: string;
+export class Photo extends FullAuditedEntity {
+    /*  @ApiProperty({ nullable: true, required: false })
+     @PrimaryGeneratedColumn('uuid')
+     id?: string; */
     @ApiProperty({ nullable: true, required: false })
     @Column({ type: 'bytea' })
     originalFile?: Buffer;
