@@ -4,6 +4,7 @@ import { Photo } from "./models/photo.entity";
 import { PhotoService } from "./service/photo.service";
 import { PhotoController } from "./controller/photo.controller";
 import { AuthModule } from "src/auth/auth.module";
+import { CoreModule } from "src/core/core.module";
 
 export const StorageEntities = [
     Photo,
@@ -13,8 +14,9 @@ export const StorageEntities = [
     imports: [
         TypeOrmModule.forFeature(StorageEntities),
         AuthModule,
+        CoreModule,
     ],
-    controllers:[
+    controllers: [
         PhotoController,
     ],
     providers: [
